Move split and merge functions to confidence.utils
